code/requirements.txt
apache-airflow[google]
pytest